requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
requests
newsapi-python
torch
transformers
nltk
scikit-learn
Jinja2
python-multipart
beautifulsoup4
nsepy